refactor(structure): move go code to /pkg/
