+++ pkg/api/apps/create.go
@@ -2,7 +2,7 @@ package apps
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/hackclub/hack-as-a-service/db"
+	"github.com/hackclub/hack-as-a-service/pkg/db"
 )
 
 func handlePOSTApp(c *gin.Context) {
