+++ pkg/dokku_deploy/etc/nginx/conf.d/00-default-vhost.conf
